pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cognitive-symbiont"
description = "Cognitive Symbiont MVP autonomous agent toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
keywords = ["autonomous", "agent", "symbiont", "rag", "streamlit"]
authors = [
  {name = "Symbiont Team"}
]
dependencies = [
  "typer==0.12.5",
  "click==8.1.7",
  "rich==13.7.1",
  "numpy==1.26.4",
  "pydantic==2.8.2",
  "orjson==3.10.7",
  "httpx==0.27.2",
  "PyYAML==6.0.2",
  "streamlit==1.38.0",
  "pyttsx3==2.90",
  "networkx==3.3",
  "requests==2.32.3",
  "PyGithub==2.4.0",
  "tenacity==8.2.3",
  "hypothesis==6.112.1",
  "prometheus-client==0.20.0",
  "feedparser==6.0.10"
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
  "pytest==8.3.2",
  "pytest-cov==5.0.0",
  "openai>=1.46.0"
]
foresight = [
  "pgmpy>=0.1.24",
  "dowhy>=0.11.1"
]

[project.scripts]
sym = "symbiont.cli:app"

[tool.setuptools]
package-dir = {"" = "."}
include-package-data = true

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.setuptools.packages.find]
where = ["."]
include = ["symbiont*","bigkit_v3alpha*","systems*","configs*","scripts*"]

[tool.setuptools.package-data]
"symbiont" = ["**/*.yaml", "**/*.yml", "**/*.sql", "**/*.json"]

